+++ tests/test_tg.py
@@ -1,4 +1,4 @@
-from ingestion.utils import format_jsonl, extract_all_messages
+from ingestion.tg import format_jsonl, extract_all_messages
 from os import getenv
 
 message_list = extract_all_messages("ingestion/sample.json")
